+++ src/core/fbs/fbs_optimizer.rs
@@ -1,14 +1,36 @@
 //! FBS Algorithm
 //!
-use super::super::AlgorithmEngine;
-use super::super::Optimizer;
-use super::super::SolverStatus;
-use super::FBSEngine;
-use super::FBSOptimizer;
-use crate::constraints;
+use crate::{
+    constraints,
+    core::{
+        fbs::fbs_engine::FBSEngine, fbs::FBSCache, AlgorithmEngine, Optimizer, Problem,
+        SolverStatus,
+    },
+};
 use std::time;
 
 const MAX_ITER: usize = 100_usize;
+
+/// Optimiser using forward-backward splitting iterations (projected gradient)
+///
+/// Note that an `FBSOptimizer` holds a reference to an instance of `FBSEngine`
+/// which needs to be created externally. A mutable reference to that `FBSEgnine`
+/// is provided to the optimizer.
+///
+/// The `FBSEngine` is supposed to be updated whenever you need to solve
+/// a different optimization problem.
+///
+///
+pub struct FBSOptimizer<'a, GradientType, ConstraintType, CostType>
+where
+    GradientType: Fn(&[f64], &mut [f64]) -> i32,
+    CostType: Fn(&[f64], &mut f64) -> i32,
+    ConstraintType: constraints::Constraint,
+{
+    fbs_engine: FBSEngine<'a, GradientType, ConstraintType, CostType>,
+    max_iter: usize,
+    max_duration: Option<time::Duration>,
+}
 
 impl<'a, GradientType, ConstraintType, CostType>
     FBSOptimizer<'a, GradientType, ConstraintType, CostType>
@@ -18,10 +40,11 @@ where
     ConstraintType: constraints::Constraint,
 {
     pub fn new(
-        fbs_engine: &'a mut FBSEngine<'a, GradientType, ConstraintType, CostType>,
+        problem: Problem<GradientType, ConstraintType, CostType>,
+        cache: &'a mut FBSCache,
     ) -> FBSOptimizer<'a, GradientType, ConstraintType, CostType> {
         FBSOptimizer {
-            fbs_engine,
+            fbs_engine: FBSEngine::new(problem, cache),
             max_iter: MAX_ITER,
             max_duration: None,
         }
@@ -33,9 +56,9 @@ where
     ///
     /// The method panics if the specified tolerance is not positive
     pub fn with_tolerance(
-        &mut self,
+        mut self,
         tolerance: f64,
-    ) -> &mut FBSOptimizer<'a, GradientType, ConstraintType, CostType> {
+    ) -> FBSOptimizer<'a, GradientType, ConstraintType, CostType> {
         assert!(tolerance > 0.0);
 
         self.fbs_engine.cache.tolerance = tolerance;
@@ -44,18 +67,18 @@ where
 
     /// Sets the maximum number of iterations
     pub fn with_max_iter(
-        &mut self,
+        mut self,
         max_iter: usize,
-    ) -> &mut FBSOptimizer<'a, GradientType, ConstraintType, CostType> {
+    ) -> FBSOptimizer<'a, GradientType, ConstraintType, CostType> {
         self.max_iter = max_iter;
         self
     }
 
     /// Sets the maximum number of iterations
     pub fn with_max_duration(
-        &mut self,
+        mut self,
         max_duration: time::Duration,
-    ) -> &mut FBSOptimizer<'a, GradientType, ConstraintType, CostType> {
+    ) -> FBSOptimizer<'a, GradientType, ConstraintType, CostType> {
         self.max_duration = Some(max_duration);
         self
     }
@@ -64,9 +87,9 @@ where
 impl<'life, GradientType, ConstraintType, CostType> Optimizer
     for FBSOptimizer<'life, GradientType, ConstraintType, CostType>
 where
-    GradientType: Fn(&[f64], &mut [f64]) -> i32,
-    CostType: Fn(&[f64], &mut f64) -> i32,
-    ConstraintType: constraints::Constraint,
+    GradientType: Fn(&[f64], &mut [f64]) -> i32 + 'life,
+    CostType: Fn(&[f64], &mut f64) -> i32 + 'life,
+    ConstraintType: constraints::Constraint + 'life,
 {
     fn solve(&mut self, u: &mut [f64]) -> SolverStatus {
         let now = time::Instant::now();
