+++ src/core/fbs/tests.rs
@@ -1,6 +1,7 @@
 use super::super::*;
 use super::*;
 use crate::constraints;
+use crate::core::fbs::fbs_engine::FBSEngine;
 use std::num::NonZeroUsize;
 
 const N_DIM: usize = 2;
@@ -21,10 +22,8 @@ fn t_solve_fbs_hard() {
     let tolerance = 1e-6;
 
     let mut fbs_cache = FBSCache::new(NonZeroUsize::new(3).unwrap(), gamma, tolerance);
-    let mut fbs_engine = FBSEngine::new(problem, &mut fbs_cache);
     let mut u = [-12., -160., 55.];
-    let mut optimizer = FBSOptimizer::new(&mut fbs_engine);
-    optimizer.max_iter = 100000;
+    let mut optimizer = FBSOptimizer::new(problem, &mut fbs_cache).with_max_iter(100000);
     let status = optimizer.solve(&mut u);
 
     println!("|fpr| = {}", status.norm_fpr());
@@ -89,9 +88,8 @@ fn t_solve_fbs() {
     let tolerance = 1e-6;
 
     let mut fbs_cache = FBSCache::new(NonZeroUsize::new(N_DIM).unwrap(), gamma, tolerance);
-    let mut fbs_engine = FBSEngine::new(problem, &mut fbs_cache);
     let mut u = [0.0; N_DIM];
-    let mut optimizer = FBSOptimizer::new(&mut fbs_engine);
+    let mut optimizer = FBSOptimizer::new(problem, &mut fbs_cache);
 
     let status = optimizer.solve(&mut u);
 
@@ -119,15 +117,12 @@ fn t_solve_fbs_many_times() {
         // The problem is surely update at every execution of NMPC
         let problem = Problem::new(box_constraints, mocks::my_gradient, mocks::my_cost);
 
-        // Construct a new Engine; this does not allocate any memory
-        let mut fbs_engine = FBSEngine::new(problem, &mut fbs_cache);
-
         // Here comes the new initial condition
         u[0] = 2.0 * _i as f64;
         u[1] = -_i as f64;
 
         // Create a new optimizer...
-        let mut optimizer = FBSOptimizer::new(&mut fbs_engine);
+        let mut optimizer = FBSOptimizer::new(problem, &mut fbs_cache);
 
         let status = optimizer.solve(&mut u);
